requirements.txt
pyyaml
langdetect
tiktoken
beautifulsoup4
lxml
numpy
matplotlib
# detoxify
# torch
